UI rebalancing, Parent-Child variants, and Unit Type system upgrade
- POS UI Rebalancing: 50/50 split, constrained resize policy for tables, and enhanced pricing mode visibility.
- Inventory Architecture: Parent-Child variant model for fractional stock deduction (e.g., 0.5kg) from master product.
- Unit Type System: Added official unit types (Kg, Liters, etc.) across database, model, and UI.
- Unit-Agnostic Labels: Updated all FXML files and controllers to replace hardcoded 'pieces' with flexible terminology.
- Window Management: Fixed responsive constraints and ensured maximized window launch without hardcoded dimensions.
- Cleanup: Removed legacy AdminController and associated FXML.

+++ src/main/java/com/pos/controller/QuickAddController.java
@@ -110,12 +110,12 @@ public class QuickAddController {
             // Parse prices safely
             BigDecimal retailPriceValue;
             BigDecimal wholesalePriceValue;
-            int stockValue;
+            double stockValue;
             
             try {
                 retailPriceValue = retailPrice.isEmpty() ? BigDecimal.ZERO : new BigDecimal(retailPrice);
                 wholesalePriceValue = wholesalePrice.isEmpty() ? BigDecimal.ZERO : new BigDecimal(wholesalePrice);
-                stockValue = stock.isEmpty() ? 0 : Integer.parseInt(stock);
+                stockValue = stock.isEmpty() ? 0 : Double.parseDouble(stock);
             } catch (NumberFormatException | ArithmeticException e) {
                 showError("Invalid price or stock value. Please enter valid numbers.");
                 return;
